requirements.txt
qrcode[pil]==7.4.2
pandas==2.1.4
reportlab==4.0.8
python-socketio[client]==5.10.0
Pillow==10.1.0